refactor(web): move workspace activate out of avatar chrome
Setup, join, and invite are chrome-free and should not import
the app avatar switcher for session activation.
