+++ apps/web/src/app/(app)/components/user-avatar/workspace-switcher.tsx
@@ -5,8 +5,7 @@ import { usePathname, useRouter } from "next/navigation";
 import { useState, useTransition } from "react";
 import { toast } from "sonner";
 
-import { updatePreferredOrganization } from "@/lib/actions/organization";
-import { authClient } from "@/lib/auth/auth.client";
+import { activateOrganizationWorkspace } from "@/lib/activate-organization-workspace";
 
 export function getAgentJobsBasePath(pathname: string): string | null {
   const jobsRouteMatch = pathname.match(/^\/agents\/([^/]+)\/jobs(?:\/.*)?$/);
@@ -25,26 +24,6 @@ export function getTaskDetailBasePath(pathname: string): string | null {
   }
 
   return "/tasks";
-}
-
-export async function activateOrganizationWorkspace(
-  organizationId: string | null,
-): Promise<void> {
-  await authClient.organization.setActive({
-    organizationId,
-  });
-
-  try {
-    const result = await updatePreferredOrganization({
-      organizationId,
-    });
-
-    if (!result.ok) {
-      console.error("Failed to persist preferred organization:", result.error);
-    }
-  } catch (error) {
-    console.error("Failed to persist preferred organization:", error);
-  }
 }
 
 interface SwitchOrganizationWorkspaceOptions {
