+++ apps/web/src/app/(setup)/setup/components/__tests__/identity-onboarding-form.test.tsx
@@ -38,7 +38,7 @@ vi.mock("@/lib/actions/workspace-gate", () => ({
     createPersonalWorkspaceActionMock(...args),
 }));
 
-vi.mock("@/app/components/user-avatar/workspace-switcher", () => ({
+vi.mock("@/lib/activate-organization-workspace", () => ({
   activateOrganizationWorkspace: (...args: unknown[]) =>
     activateOrganizationWorkspaceMock(...args),
 }));
